Fix minor issues
Add to lines in playRound() function to trim and make lowercase the
playerSelection and computerSelection values. Also modified the default
block from the switch statement that sets "result" variable to set it
to undefined when the playerSelection value is not a valid options.

With that change, a new condition was added to the returned function
from game() function to handle the possibility that the player passes an
invalid string.