Replace game() function with an object
game() function was in charge of managing the game logic but wasn't
very flexible with private variable, so it was replace with an object
using the module pattern. This new game object functions the exact same
way as game() function with the exception that now there is more control
over private variables.