Update design
Place #game-announcer on top of .game-display and updated the selectors
of .weapon-button to show a different background color when the button
it's enabled and to be opaque when it's disabled. Also added max-width
for body to not be more than 900px wide and a grey background color to
be able to see clearly the paper image.